+++ src/BlazorServer.BackOffice/Models/Article/SearchArticlesResponse.cs
@@ -2,7 +2,7 @@
 {
     public class SearchArticlesResponse
     {
-        public SearchArticlesResponse(Guid id, string name, string categoryName, double basePrice)
+        public SearchArticlesResponse(Guid id, string name, string categoryName, decimal basePrice)
         {
             Id = id;
             Name = name;
@@ -13,6 +13,6 @@
         public Guid Id { get; set; }
         public string Name { get; set; }
         public string CategoryName { get; set; }
-        public double BasePrice { get; set; }
+        public decimal BasePrice { get; set; }
     }
 }
